requirements.txt
openai
faiss-cpu
tqdm
python-dotenv
loguru